preparing 64bit build
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@4609 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSLaneChanger.cpp
@@ -702,7 +702,7 @@ MSLaneChanger::findTarget()
         }
         if (candiOnAllowed(ce)) {
 
-            unsigned int dist = abs(distance(myCandi, ce));
+            unsigned int dist = abs((int)distance(myCandi, ce));
             assert(dist > 0);
             assert(dist < myChanger.size());
             if (dist < minDist) {
